@@ -23,54 +25,59 @@
     "dependencies": {
         "@babel/core": "^7.4.4",
         "@mapbox/polyline": "^1.0.0",
+        "@react-native-community/geolocation": "^2.0.2",
+        "@react-native-community/netinfo": "^2.0.2",
         "babel-polyfill": "^6.26.0",
         "eslint-staged-files": "1.0.0",
         "firebase": "^5.0.3",
         "grpc": "^1.20.2",
         "haversine": "^1.1.0",
-        "react": "16.6.3",
-        "react-native": "0.58.6",
+        "react": "16.9.0",
+        "react-native": "0.61.5",
         "react-native-actionsheet": "^2.4.2",
         "react-native-activity-recognition": "^3.2.0",
         "react-native-android-location-services-dialog-box": "^2.4.4",
         "react-native-background-timer": "^2.0.1",
         "react-native-email-link": "^1.4.0",
-        "react-native-fbsdk": "^0.8.0",
+        "react-native-fbsdk": "1.0.2",
         "react-native-fetch-blob": "^0.10.8",
-        "react-native-firebase": "^5.5.3",
+        "react-native-firebase": "^5.6.0",
         "react-native-geocoding": "^0.3.0",
         "react-native-google-places": "^2.5.0",
         "react-native-google-signin": "^2.0.0",
-        "react-native-image-picker": "^0.26.10",
-        "react-native-keyboard-aware-scroll-view": "^0.5.0",
+        "react-native-image-picker": "^1.1.0",
+        "react-native-keyboard-aware-scroll-view": "^0.9.0",
         "react-native-maps": "^0.21.0",
         "react-native-modal": "^11.1.0",
         "react-native-modal-translucent": "^2.0.0",
-        "react-native-offline": "^3.14.1",
+        "react-native-offline": "^5.1.0",
         "react-native-picker": "^4.3.6",
         "react-native-router-flux": "^3.43.0",
-        "react-native-scrollable-tab-view": "^0.8.0",
+        "react-native-screens": "^2.2.0",
+        "react-native-scrollable-tab-view": "^1.0.0",
         "react-native-simple-auth": "^2.3.0",
         "react-native-simple-toast": "0.0.8",
         "react-native-snap-carousel": "^3.8.0",
-        "react-native-splash-screen": "^3.0.6",
-        "react-native-swiper": "^1.5.13",
-        "react-native-timeline-listview": "^0.2.3",
-        "react-native-vector-icons": "^4.6.0",
-        "react-redux": "^5.0.7",
+        "react-native-splash-screen": "3.1.1",
+        "react-native-swiper": "^1.6.0-nightly.5",
+        "react-native-timeline-flatlist": "^0.6.5",
+        "react-native-vector-icons": "^6.6.0",
+        "react-redux": "^7.1.3",
         "realm": "^2.4.0",
         "redux": "^4.0.0",
         "redux-logger": "^3.0.6",
         "redux-thunk": "^2.2.0"
     },
     "devDependencies": {
-        "babel-core": "^7.0.0-bridge.0",
-        "babel-jest": "22.4.3",
+        "@babel/core": "^7.6.2",
+        "@babel/runtime": "^7.6.2",
+        "@react-native-community/eslint-config": "^0.0.5",
+        "babel-jest": "^24.9.0",
         "babel-polyfill": "^6.26.0",
         "enzyme": "^3.8.0",
         "enzyme-adapter-react-16": "^1.7.1",
         "enzyme-to-json": "^3.3.5",
-        "eslint": "5.15.0",
+        "eslint": "^6.5.1",
         "eslint-config-airbnb": "17.1.0",
         "eslint-config-prettier": "4.1.0",
         "eslint-plugin-import": "2.16.0",
